refactor(sidebar): H2.3 — extract useSidebarNavDnd hook
